init: adding skeleton for environment modules
This currently works to parse environment modules
(and not LMOD). We can add the latter if needed. It is a
fairly basic parsing that represents the top level module
(name and version) as nodes, and any binary/libs discovered
within the paths derived as children "binary" and "library."
For my testing case I do not have any because I used a dummy
repository of module files for which I do not actually have
the software for.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>